+++ static/js/components/followup_request/FollowupRequestForm.tsx
@@ -9,7 +9,7 @@ import Select from "@mui/material/Select";
 import { makeStyles } from "tss-react/mui";
 import Form, { Templates as MuiTemplates } from "@rjsf/mui";
 import validator from "@rjsf/validator-ajv8";
-import { asNumber, getUiOptions } from "@rjsf/utils";
+import { getUiOptions } from "@rjsf/utils";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
 import DialogTitle from "@mui/material/DialogTitle";
@@ -32,6 +32,7 @@ import {
   isSomeActiveRangeOrNoRange,
   rangeIsActive,
 } from "../allocation/AllocationTable";
+import { localeSafeFields } from "./LocaleSafeNumberField";
 
 const useStyles = makeStyles()(() => ({
   marginTop: {
@@ -102,50 +103,6 @@ const FollowupFieldTemplate = (props: any) => {
   return <MuiFieldTemplate {...props} />;
 };
 
-// rjsf-core's stock NumberField re-formats numeric values using the OS/browser
-// locale's decimal separator (e.g. "." -> ",") before writing them back into
-// the input. Native <input type="number"> elements always require "."
-// regardless of locale, so on comma-locale systems the browser rejects the
-// reformatted value ("The specified value "2,8" cannot be parsed") and the
-// field renders blank. This is the same field, minus that reformatting step.
-const numberTrailingCharMatcherWithPrefix = /\.([0-9]*0)*$/;
-const numberTrailingCharMatcher = /[0.]0*$/;
-const FollowupNumberField = (props: any) => {
-  const { formData, onChange, registry } = props;
-  const [lastValue, setLastValue] = useState(formData);
-  const { StringField } = registry.fields;
-
-  let value = formData;
-  if (typeof lastValue === "string" && typeof value === "number") {
-    const escapedValue = String(value).replace(".", "\\.");
-    const re = new RegExp(`^(${escapedValue})?\\.?0*$`);
-    if (lastValue.match(re)) {
-      value = lastValue;
-    }
-  }
-
-  const handleChange = (
-    newValue: any,
-    path: any,
-    errorSchema: any,
-    id: any,
-  ) => {
-    setLastValue(newValue);
-    const normalizedValue =
-      typeof newValue === "string" && newValue.startsWith(".")
-        ? `0${newValue}`
-        : newValue;
-    const processed =
-      typeof normalizedValue === "string" &&
-      numberTrailingCharMatcherWithPrefix.exec(normalizedValue)
-        ? asNumber(normalizedValue.replace(numberTrailingCharMatcher, ""))
-        : asNumber(normalizedValue);
-    onChange(processed, path, errorSchema, id);
-  };
-
-  return <StringField {...props} formData={value} onChange={handleChange} />;
-};
-
 // Stable reference: a new object literal here would make rjsf rebuild its
 // registry on every keystroke, resetting fields' local state (e.g. NumberField's
 // in-progress-decimal cache), which erased values like "2.5" while typing.
@@ -153,7 +110,6 @@ const followupTemplates = {
   BaseInputTemplate: FollowupBaseInputTemplate,
   FieldTemplate: FollowupFieldTemplate,
 };
-const followupFields = { NumberField: FollowupNumberField };
 
 interface FollowupRequestFormProps {
   obj_id: string;
@@ -581,7 +537,7 @@ const FollowupRequestForm = ({
             validator={validator}
             uiSchema={uiSchema}
             templates={followupTemplates}
-            fields={followupFields}
+            fields={localeSafeFields}
             customValidate={validate}
             onSubmit={handleSubmit as any}
             disabled={isSubmitting}
